+++ ecs/query_test.go
@@ -36,6 +36,7 @@ func TestQuery(t *testing.T) {
 		cnt++
 	}
 	assert.Equal(t, 2, cnt)
+	assert.Equal(t, 2, q.Count())
 
 	q = w.Query(posID)
 	cnt = 0
@@ -47,6 +48,7 @@ func TestQuery(t *testing.T) {
 		cnt++
 	}
 	assert.Equal(t, 3, cnt)
+	assert.Equal(t, 3, q.Count())
 
 	q = w.Query(rotID)
 	cnt = 0
@@ -60,6 +62,7 @@ func TestQuery(t *testing.T) {
 		cnt++
 	}
 	assert.Equal(t, 4, cnt)
+	assert.Equal(t, 4, q.Count())
 
 	assert.Panics(t, func() { q.Next() })
 	assert.Panics(t, func() { q.Entity() })